Changed from unsafe casting to type checking

+++ comet/src/commands/text/MemeCommand.ts
@@ -1,4 +1,4 @@
-import {Message, TextChannel} from "discord.js";
+import { Message, TextChannel } from "discord.js";
 import Elia from "../../Elia";
 import Command from "../Command";
 import { CommandTypeEnum } from "../CommandTypeEnum";
@@ -16,8 +16,8 @@ export default class MemeCommand extends Command {
                 message.guild.id
             );
             if (channelID) {
-                const channel = await message.client.channels.fetch(channelID) as TextChannel;
-                if (channel) {
+                const channel = await message.client.channels.fetch(channelID);
+                if (channel && channel instanceof TextChannel) {
                     const messageText = args.join(" ");
                     channel.send(`${message.author.toString()} ` + messageText);
                     message.delete();
